Move InfoButton to com.trafi.ui.component package.
