+++ android/app/src/androidTest/java/com/trafi/example/TestPreviews.kt
@@ -1,10 +1,10 @@
 package com.trafi.example
 
 import androidx.compose.runtime.Composable
-import com.trafi.ui.InfoButtonDisabledPreview
-import com.trafi.ui.InfoButtonLongTextPreview
-import com.trafi.ui.InfoButtonStyledDarkPreview
-import com.trafi.ui.InfoButtonThemedPreview
+import com.trafi.ui.component.InfoButtonDisabledPreview
+import com.trafi.ui.component.InfoButtonLongTextPreview
+import com.trafi.ui.component.InfoButtonStyledDarkPreview
+import com.trafi.ui.component.InfoButtonThemedPreview
 import com.trafi.ui.component.BadgePreview
 import com.trafi.ui.component.BadgePreviewWithSubbadge
 import com.trafi.ui.component.BadgeWithoutIconPreview
